settings are being saved now BUT
when the username gets changed: the person changing the username has an extra user in the members list, which is their old name and the update message and new name is also only displayed to them. name colours are also only visible to the current user. the other users can see the new name. idk i am still working on it but i wanted to push for now.